[TL42-252] refactor: refresh issue 해결 시도
