+++ front-end/src/UI/Templates/ChatMessage/index.tsx
@@ -1,205 +1,59 @@
-import React, { useCallback, useEffect } from 'react';
-import { useQuery, useQueryClient } from '@tanstack/react-query';
-import { useNavigate, useParams } from 'react-router-dom';
-import { FormikHelpers } from 'formik';
+import React from 'react';
 import axios from 'axios';
-import useMe from '../../../Hooks/useMe';
-import useMessage from '../../../Hooks/useMessage';
-import { useSocket } from '../../../Hooks/useSocket';
+import { useQueryClient } from '@tanstack/react-query';
+import { useNavigate, useParams } from 'react-router-dom';
 import useWarningDialog from '../../../Hooks/useWarningDialog';
-import CHAT_USERS_GET from '../../../Queries/ChatUsers/All';
-import ChatModal from '../../Organisms/ChatModal';
-import SocketEventName from '../../../WebSockets/dto/constants/socket.events.enum';
-import ChatJoinNotifyProps from '../../../WebSockets/dto/res/chat.join.notify.dto';
-import ChatMemberRole from '../../../Props/ChatMemberRole';
-import ChatLeaveNotifyProps from '../../../WebSockets/dto/res/chat.leave.notify.dto';
-import ChatMessageProps from '../../../WebSockets/dto/res/chat.message.notify.dto';
-import ChatUserProps from '../../../Props/ChatUserProps';
-import ROOM_GET from '../../../Queries/Channels/Room';
-import ChatUpdateProtectionNotifyProps from '../../../WebSockets/dto/res/chat.update.protection.notify.dto';
-import ChannelType from '../../../Props/ChannelType';
+import ChatMessageContent from '../../Organisms/ChatMessageContent';
+import useMe from '../../../Hooks/useMe';
 
 export default function ChatMessage() {
-  const { dispatchRoomInfo, dispatchChat, insertRoomMember, deleteRoomMember } =
-    useMessage();
-  const navigate = useNavigate();
-  const { nickname } = useMe();
   const { id } = useParams();
-  const chatRoomId = id ?? '';
-  const { isLoading, error, data } = useQuery(CHAT_USERS_GET(chatRoomId));
-  const { setError, WarningDialogComponent } = useWarningDialog(() =>
-    navigate(-1),
-  );
-  const { state } = useSocket();
+  const roomId = React.useMemo(() => id ?? '', [id]);
   const queryClient = useQueryClient();
-  const {
-    error: roomInfoError,
-    isLoading: roomInfoLoading,
-    data: roomData,
-  } = useQuery(ROOM_GET(chatRoomId));
+  const navigate = useNavigate();
+  const { setError, WarningDialogComponent } = useWarningDialog(() => {
+    navigate(-1);
+  });
+  const [isSuccess, setSuccess] = React.useState(false);
+  const { id: myId } = useMe();
 
-  const onSubmitHandler = useCallback(
-    (
-      values: { message: string },
-      helper: FormikHelpers<{ message: string }>,
-    ) => {
-      const { message } = values;
-      if (message.length === 0) return;
-      axios
-        .post(`/chat/send/${chatRoomId}`, { content: message })
-        .then(() => {
-          dispatchChat(nickname, message);
-          helper.resetForm();
-        })
-        .catch((err) => {
-          if (err.response) {
-            setError({
-              headerMessage: '오류 발생',
-              bodyMessage: err.response.data.message,
-            });
-          } else {
-            setError({
-              headerMessage: '오류 발생',
-              bodyMessage: err.message,
-            });
-          }
-        });
-    },
-    [chatRoomId, dispatchChat, nickname, setError],
-  );
-
-  const leaveModalHandler = () => {
-    if (chatRoomId === '') return;
+  const password = React.useMemo(() => {
+    const localPassword = localStorage.getItem(roomId);
+    if (localPassword) return { password: localPassword };
+    return undefined;
+  }, [roomId]);
+  React.useEffect(() => {
+    if (!roomId) {
+      setError({
+        headerMessage: '채팅 입장 실패',
+        bodyMessage: '잘못된 접근 입니다.',
+      });
+      return;
+    }
     axios
-      .delete(`/chat/leave/${chatRoomId}`)
-      .then(() => queryClient.invalidateQueries(['channels']))
+      .post(`/chat/join/${roomId}`, password)
+      .then(() => {
+        queryClient.invalidateQueries(['channels']);
+        setSuccess(true);
+      })
       .catch((err) => {
         if (err.response) {
           setError({
-            headerMessage: '오류 발생',
+            headerMessage: '채팅 입장 실패',
             bodyMessage: err.response.data.message,
           });
         } else {
           setError({
-            headerMessage: '오류 발생',
+            headerMessage: '채팅 입장 실패',
             bodyMessage: err.message,
           });
         }
       });
-  };
-
-  useEffect(() => {
-    if (isLoading || roomInfoLoading) return;
-    if (!chatRoomId || error || roomInfoError) {
-      setError({
-        headerMessage: '오류 발생',
-        bodyMessage: error
-          ? '서버와의 통신에 실패했습니다.'
-          : '정상적인 접근이 아닙니다.',
-      });
-    }
-  }, [chatRoomId, error, isLoading, roomInfoLoading, roomInfoError, setError]);
-
-  useEffect(() => {
-    axios
-      .get(`/chat/users/${chatRoomId}`)
-      .then((response: { data: ChatUserProps[] }) => {
-        const result = response.data;
-        result.forEach((member) => {
-          insertRoomMember({
-            userId: member.user.id,
-            name: member.user.nickname,
-            profileImg: `${process.env.REACT_APP_API_HOST}/${member.user.profileImg}`,
-            role: member.role,
-            muted: false,
-            blocked: false,
-          });
-        });
-      });
-  }, [insertRoomMember, chatRoomId, nickname]);
-
-  useEffect(() => {
-    dispatchRoomInfo({
-      roomType: roomData?.type ?? ChannelType.PUBLIC,
-      channelName: roomData?.name ?? '',
-    });
-  }, [dispatchRoomInfo, roomData]);
-
-  useEffect(() => {
-    state.socket?.on(
-      SocketEventName.CHAT_JOIN_NOTIFY,
-      (dto: ChatJoinNotifyProps) => {
-        insertRoomMember({
-          profileImg: `${process.env.REACT_APP_API_HOST}/${dto.profileImg}`,
-          userId: dto.id,
-          name: dto.nickname,
-          role: ChatMemberRole.MEMBER,
-          muted: false,
-          blocked: false,
-        });
-        if (dto.nickname === nickname) {
-          data
-            ?.filter((m) => m.user.nickname !== nickname)
-            .forEach((member) => {
-              insertRoomMember({
-                userId: member.user.id,
-                name: member.user.nickname,
-                profileImg: `${process.env.REACT_APP_API_HOST}/${member.user.profileImg}`,
-                role: member.role,
-                muted: false,
-                blocked: false,
-              });
-            });
-        }
-      },
-    );
-    state.socket?.on(
-      SocketEventName.CHAT_LEAVE_NOTIFY,
-      (dto: ChatLeaveNotifyProps) => {
-        deleteRoomMember(dto.nickname);
-      },
-    );
-    state.socket?.on(
-      SocketEventName.CHAT_MESSAGE_NOTIFY,
-      (dto: ChatMessageProps) => {
-        if (dto.nickname !== nickname) {
-          dispatchChat(dto.nickname, dto.content);
-        }
-      },
-    );
-
-    state.socket?.on(
-      SocketEventName.CHAT_UPDATE_PROTECTION_NOTIFY,
-      (dto: ChatUpdateProtectionNotifyProps) => {
-        dispatchRoomInfo({
-          roomType: dto.status ? ChannelType.PUBLIC : ChannelType.PROTECT,
-          channelName: roomData?.name ?? '',
-        });
-      },
-    );
-    return () => {
-      state.socket?.off(SocketEventName.CHAT_JOIN_NOTIFY);
-      state.socket?.off(SocketEventName.CHAT_LEAVE_NOTIFY);
-      state.socket?.off(SocketEventName.CHAT_MESSAGE_NOTIFY);
-      state.socket?.off(SocketEventName.CHAT_UPDATE_PROTECTION_NOTIFY);
-    };
-  }, [
-    data,
-    nickname,
-    state.socket,
-    insertRoomMember,
-    deleteRoomMember,
-    dispatchChat,
-    dispatchRoomInfo,
-    roomData?.name,
-  ]);
+  }, [roomId, navigate, password, queryClient, setError, setSuccess, myId]);
 
   return (
     <>
-      <ChatModal
-        onSubmitHandler={onSubmitHandler}
-        leaveModalHandler={leaveModalHandler}
-      />
+      {isSuccess && <ChatMessageContent setError={setError} />}
       {WarningDialogComponent}
     </>
   );
